Combinatorial sbol

+++ src/main/java/knox/spring/data/neo4j/operations/ANDOperator.java
@@ -1,12 +1,18 @@
 package knox.spring.data.neo4j.operations;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import knox.spring.data.neo4j.domain.Edge;
 import knox.spring.data.neo4j.domain.NodeSpace;
 
 public class ANDOperator {
+	
+	private static final Logger LOG = LoggerFactory.getLogger(NodeSpace.class);
 	
 	public static void apply(List<NodeSpace> inputSpaces, NodeSpace outputSpace, 
 			int tolerance, boolean isComplete, Set<String> roles) {
@@ -25,13 +31,17 @@ public class ANDOperator {
 				product.getSpace().deleteBlankEdges(blankEdges.get(j));
 			}
 		}
+		
+		if (product.getSpace().hasNodes()) {
+			Union union = new Union(product.getSpace());
 
-		Union union = new Union(product.getSpace());
+			Set<Edge> blankEdges = union.apply();
 
-		Set<Edge> blankEdges = union.apply();
+			union.getSpace().deleteBlankEdges(blankEdges);
 
-		union.getSpace().deleteBlankEdges(blankEdges);
-
-		outputSpace.shallowCopyNodeSpace(union.getSpace());
+			outputSpace.shallowCopyNodeSpace(union.getSpace());
+		} else {
+			outputSpace.shallowCopyNodeSpace(new NodeSpace(new ArrayList<String>(), new ArrayList<String>()));
+		}
 	}
 }
